Handle corner case for evict with rfe NULL
Under special circumstances, for BGPaaS, we might try to evict a reverse flow
(from Controller side) which points to a forward flow which has been already
deleted. This can happen when two different flows from the VNF are trapped to
agent, combined with the Controller resetting/closing the TCP connection.

Closes-Jira-bug: CEM-18166

Signed-off-by: Ciprian Barbu <cbarbu@juniper.net>
Change-Id: I3a9ddcdeee9dc2c6a728c1b16fd87175e2b6ad71