DPDK crashes when memcpy to a NULL ptr in virtio
Description:
It was found that during geust VM failover, guest VM passes the invalid guest physical address in desc table. When DPDK decodes this guest physical address from VRING and it deems it is invalid memory and translates it to NULL instead of valid vhost virtual address. When DPDK sends packets to VM, it doesn't check the validity of this vhost virtual address before memcpy to it, so a NULL ptr can cause seg fault.

Fix:
Sanity check to ensure the address is not NULL before memcpy to VRING.

Closes-Jira-Bug: CEM-20290
Signed-off-by: Nan Wang <nanw@juniper.net>
Change-Id: Ibdf6013b6672605cab8ca3e24d970764aef0fc05