DT-NIMS hotfix: histograms and logging vrouter code changes.
This patch is made to support the changes made in the following dpdk commit:
https://review.opencontrail.org/c/Juniper/contrail-dpdk/+/63682.
Here, the dpdkinfo tool is being used to fetch and display the histograms which are
populated by the changes made in the above dpdk commit.

Commmand used to display the histograms: dpdkinfo -l all
In the above commnad, there is a hidden "dpdkinfo -l clear" option which can be
used during debugging. This option will reset all the histogram statistics.

Note: This patch is a hot-fix to address the DT issue CEM-21858 and the patch will
be reverted when a build becomes available after the merge. The patch is being
reverted since this monitoring tool is specifically designed for the DT environment
and has some shortcomings such as it is only valid for a setup functioning in fast
LACP mode and it is not production ready for all contrail environments. A JIRA story
https://contrail-jws.atlassian.net/browse/CN2-2822 has been created to address this
issue and make this tool a full fledged monitoring framework for all contrail
environments for future releases.

Partial-Jira-Bug: CEM-21858
Signed-off-by: Vinay Agrawal <vagrawal@juniper.net>
Change-Id: I82b3b47245d97595b33947bc17a64b921d21d352
Depends-On: Ic042d839eb44bb6334dc92f5e8e237253feefa1a