Fix for netlink lcore deadlock
In the stress testing of VM reboots, there was a deadlock between the
uvhost lcore (lcore-2) and netlink lcore (lcore-9). The VMs were enabled
with multiqueue virtio.

Scenario 1:
The uvhost core sends back to back post events to netlink lcore to setup
the queue in the function vr_dpdk_virtio_tx_queue_enable_disable(). At the
same time, vif delete triggers the function vr_dpdk_lcore_if_unschedule()
in the netlink thread which waits for all lcores to finish executing their
current commands (including uvhost thread). So, there will be a case where
both uvhost and netlink threads are waiting on each other resulting in
a deadlock.

Fix to scenario 1:
The fix is to handle any outstanding commands for the "self" lcore before
it waits for other lcores to complete executing their commands. The fix
is in the function vr_dpdk_lcore_cmd_wait()

Scenario 2:
It is the same as scenario 1 for uvhost thread. But netlink thread executes
the function vr_delay_op() and waits for RCU quiescent state of all threads.
But uvhost thread cannot enter quiescent state since it is waiting for the
netlink thread to complete execution. So, again there is a deadlock.

Fix for scenario 2:
The fix is to make the uvhost thread enter the quiescent state using
rcu_thread_offline() call before posting a command in the function
vr_dpdk_lcore_cmd_post(). After posting, it can resume the read critical
section using rcu_thread_online() call.

Closes-Jira-Bug: CEM-24855
Signed-off-By: Kiran <kirankn@juniper.net>
Change-Id: Ie0d7853e5bc6679f9800cd18365bba67e1054ec0