Fix Title: vrmemstats util needs to be enhances to support all objects , currently many objects are missing.
Problem statement:
Memleaks are seen in DT NIMS Lab. The vrmemstats util was novered with all the counters used as part of vr_malloc, vr_zalloc and vr_free. All the missing objects are to be made part of vrmemstats util .

Proposed fix:
Identified the missing Objects/Counters . In total there are 80 objects. vrmemstats had 68 counters in the existing implementation. Eveluation is done for the 80 objects , added 10 objects . Excluded two objects in vrmemstats as they are used for indexing in the response code.

Testing done: yes

Closes-Jira-Bug: CEM-25652
Change-Id: Iebc141575bf29b0ee3b1aa6656d170d51ae489a1