Fix for memory leak, fd leak and vrouter crash due to oversized dump buffer issues in vRotuer code:
1. In the vrouter code, it was found that 3 objects which call the vr_malloc and vr_zalloc to
create memory are not freeing the memory after use. This causes a memory leak on the huge
pages used by vrouter which can eventually lead to different problems such as the intermittent
netlink issue as observed in DT setups, or the allocation of multiple hugepages which are not
cleared. Predominantly, all of this leaks are in the code paths which are related to the
vrouter utilities. Following are the 3 objects which were found to be causing the memory leaks:

a. VR_INFO_REQ_OBJECT: An object of this type is allocated when the utility commands in dpdkinfo
are triggered and leaks in a specific case when the response message size is less than 4K or
when dpdkinfo -x <all,slave1,slave2> option is triggered for any size of response message. Here,
one of the cases is that the VI_PRINTF macro has a return in error case which is not handled
properly in the display_xstats hence causing the allocations done in the display_xstats to never
be freed.
b. VR_LOG_TYPES_OBJECT: An object of this type is allocated when the utility command vrouter --info
or vrouter --get-enabled-log-types is triggered and the code tries to fetch the enabled log types.
c. VR_INTERFACE_BOND_OBJECT: An object of this type is allocated when the utility command
vif --get 0 --get-drop-stats is triggered. This leak was introduced, probably by mistake, here:
https://review.opencontrail.org/c/Juniper/contrail-vrouter/+/45600

2. The fd leak is in the dpdkinfo --app utility where the fds for the nic and bond are opened but
never closed after use resulting in many stale open fds on the compute which evenetually fill max
number of open fds allowed on the compute and hence no more fds can be created resulting in many
problems. This issue was triggered when we ran dpdkinfo --app in a loop.

3. The vrouter crash due to the oversized dump buffer gets triggered when ./dpdkinfo -xall returns
a buffer of length 4096 which is probably also the size of the sandesh object to keep the cache
aligned. In the vr_info_req_process which is the parent of all dpdkinfo functions, a NULL character
is added at the end of string which is supposed to be dumped to the sandesh. This addition of NULL
character causes the size to go to 4097 which causes a cache miss alignment which eventually causes
a seg fault.

Closes-Jira-Bug: CEM-25760
Signed-off-by: Vinay Agrawal <vagrawal@juniper.net>
Change-Id: I0082ec1f0419823b30ea44f5750163c76996855a