Fix for memory leak issues in vRotuer code:
In the vrouter code, it was found that 3 objects which call the vr_malloc and vr_zalloc to
create memory are not freeing the memory after use. This causes a memory leak on the huge
pages used by vrouter which can eventually lead to different problems such as the intermittent
netlink issue as observed in DT setups, or the allocation of multiple hugepages which are not
cleared. Predominantly, all of this leaks are in the code paths which are related to the
vrouter utilities. Following are the 3 objects which were found to be causing the memory leaks:

1. VR_INFO_REQ_OBJECT: An object of this type is allocated when the utility commands in dpdkinfo
are triggered and leaks in a specific case when the response message size is less than 4K.
2. VR_LOG_TYPES_OBJECT: An object of this type is allocated when the utility command vrouter --info
or vrouter --get-enabled-log-types is triggered and the code tries to fetch the enabled log types.
3. VR_INTERFACE_BOND_OBJECT: An object of this type is allocated when the utility command
vif --get 0 --get-drop-stats is triggered. This leak was introduced, probably by mistake, here:
https://review.opencontrail.org/c/Juniper/contrail-vrouter/+/45600

Closes-Jira-Bug: CEM-25760
Signed-off-by: Vinay Agrawal <vagrawal@juniper.net>
Change-Id: I0082ec1f0419823b30ea44f5750163c76996855a